+++ packages/gatsby-plugin-image/src/components/static-image.server.tsx
@@ -1,12 +1,11 @@
 import React, { FunctionComponent } from "react"
-import { splitProps, StaticImageProps } from "../utils"
-import { FluidObject, FixedObject } from "gatsby-image"
+import { StaticImageProps } from "../utils"
 import { GatsbyImage as GatsbyImageServer } from "./gatsby-image.server"
-import { GatsbyImageProps } from "./gatsby-image.browser"
+import { GatsbyImageProps, ISharpGatsbyImageData } from "./gatsby-image.browser"
 
 // These values are added by Babel. Do not add them manually
 interface IPrivateProps {
-  __imageData?: FluidObject & FixedObject
+  __imageData?: ISharpGatsbyImageData
   __error?: string
 }
 
@@ -22,53 +21,9 @@ export function _getStaticImage(
     if (__error) {
       console.warn(__error)
     }
-    const { gatsbyImageProps, layout } = splitProps({ src, ...props })
+
     if (imageData) {
-      const isResponsive = layout !== `fixed`
-      const childProps: Pick<
-        GatsbyImageProps,
-        "layout" | "width" | "height" | "images" | "placeholder"
-      > = {
-        layout,
-        placeholder: null,
-        width: imageData.width,
-        height: imageData.height,
-        images: {
-          fallback: {
-            src: imageData.src,
-            srcSet: imageData.srcSet,
-            sizes: isResponsive ? imageData.sizes : undefined,
-          },
-          sources: [],
-        },
-      }
-
-      if (layout === `responsive`) {
-        childProps.width = 1
-        childProps.height = imageData.aspectRatio
-      }
-
-      if (layout === `intrinsic`) {
-        childProps.width = imageData.width
-        childProps.height = imageData.height
-      }
-
-      const placeholder = imageData.tracedSVG || imageData.base64
-
-      if (placeholder) {
-        childProps.placeholder = {
-          fallback: placeholder,
-        }
-      }
-
-      if (imageData.srcWebp) {
-        childProps.images.sources.push({
-          srcSet: imageData.srcSetWebp,
-          type: `image/webp`,
-          sizes: isResponsive ? imageData.sizes : undefined,
-        })
-      }
-      return <GatsbyImage {...gatsbyImageProps} {...childProps} />
+      return <GatsbyImage image={imageData} {...props} />
     }
     console.warn(`Image not loaded`, src)
     if (!__error && process.env.NODE_ENV === `development`) {