+++ packages/gatsby-plugin-mdx/__tests__/gatsby-node.js
@@ -7,11 +7,12 @@ describe(`pluginOptionsSchema`, () => {
       `"extensions[0]" must be a string`,
       `"extensions[1]" must be a string`,
       `"extensions[2]" must be a string`,
-      `"defaultLayout" must be of type object`,
+      `"defaultLayouts" must be of type object`,
       `"gatsbyRemarkPlugins[0]" does not match any of the allowed types`,
       `"gatsbyRemarkPlugins[1]" does not match any of the allowed types`,
       `"remarkPlugins" must be an array`,
       `"rehypePlugins" must be an array`,
+      `"plugins[0]" does not match any of the allowed types`,
       `"mediaTypes[0]" must be a string`,
       `"mediaTypes[1]" must be a string`,
       `"shouldBlockNodeFromTransformation" must have an arity lesser or equal to 1`,
@@ -19,10 +20,11 @@ describe(`pluginOptionsSchema`, () => {
 
     const { errors } = await testPluginOptionsSchema(pluginOptionsSchema, {
       extensions: [1, 2, 3],
-      defaultLayout: `this should be an object`,
+      defaultLayouts: `this should be an object`,
       gatsbyRemarkPlugins: [1, { not: `existing prop` }, `valid one`],
       remarkPlugins: `this should be an array of object`,
       rehypePlugins: `this should be an array of object`,
+      plugins: [2],
       mediaTypes: [1, 2],
       shouldBlockNodeFromTransformation: (wrong, number) => null,
     })
@@ -33,7 +35,7 @@ describe(`pluginOptionsSchema`, () => {
   it(`should validate the schema`, async () => {
     const { isValid } = await testPluginOptionsSchema(pluginOptionsSchema, {
       extensions: [`.mdx`, `.mdxx`],
-      defaultLayout: {
+      defaultLayouts: {
         posts: `../post-layout.js`,
         default: `../default-layout.js`,
       },
@@ -50,6 +52,7 @@ describe(`pluginOptionsSchema`, () => {
         require(`../gatsby-node.js`),
         [require(`../gatsby-node.js`), { target: false }],
       ],
+      plugins: [{ resolve: `remark-autolink-plugin` }],
       rehypePlugins: [
         require(`../gatsby-node.js`),
         [require(`../gatsby-node.js`), { behavior: `wrap` }],