+++ packages/gatsby/src/commands/plugin-add.ts
@@ -1,0 +1,6 @@
+export default async function run({ plugins }: IProgram): Promise<void> {
+  if (!plugins?.length) {
+    console.log(`Please specify a plugin to install`)
+  }
+  console.log(plugins)
+}