+++ packages/gatsby/src/utils/__tests__/find-page-by-path.ts
@@ -50,7 +50,7 @@ describe(`findPageByPath`, () => {
   function assertGetterWithSlashMatrix(
     pathWithoutTrailingOrLeadingSlashes: string,
     exactPath: string
-  ): any {
+  ): void {
     const withTrailingSlash = pathWithoutTrailingOrLeadingSlashes + `/`
     const withLeadingSlash = `/` + pathWithoutTrailingOrLeadingSlashes
     const withLeadingAndTrailingSlash =